Safari-609.1.11.1/LayoutTests/editing/deleting/smart-delete-across-editable-boundaries.html

<!DOCTYPE html>
<html>
<body>
<script src="../../resources/dump-as-markup.js"></script>
<div contenteditable="true" id="div">foo<span contenteditable="false"> bar</span></div>
<script>
if (window.internals)
    internals.settings.setEditingBehavior('mac');
var div = document.getElementById("div");
var sel = window.getSelection();
sel.setPosition(div, 0);
sel.modify("extend", "forward", "word");
document.execCommand("Delete");
Markup.description("This tests for a bug where expansion for smart delete would not consider editable boundaries. Only 'foo' should be deleted. You should see ' bar'. <radr://problem/5390681>");
Markup.dump(div);
</script>
</body>
</html>
