LayoutTests/editing/deleting/smart-delete-across-editable-boundaries-expected.txt

This tests for a bug where expansion for smart delete would not consider editable boundaries. Only 'foo' should be deleted. You should see ' bar'. <radr://problem/5390681>
| <#selection-caret>
| <span>
|   contenteditable="false"
|   " bar"
